perf(ci): Trigger lint workflow conditionally
[Before]
This workflow would run on each push to main.
This may result in redundant workflow execution where relevant
modules to be linted have not been modified.

[After]
Define paths to conditionally trigger the workflow.
Prevent unnecessary runs for lint job, and only run on updates
to Python modules located in the specific paths defined.

Signed-off-by: Eugene Mwangi <mwangi.em37@gmail.com>